+++ api/src/main/scala/pl/touk/nussknacker/engine/api/process/package.scala
@@ -1,0 +1,5 @@
+package pl.touk.nussknacker.engine.api
+
+package object process {
+  @deprecated("Renamed to ComponentUsage", "1.3") type RunMode = ComponentUsage
+}