+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/compiler/FlinkProcessCompiler.scala
@@ -4,7 +4,7 @@ import com.typesafe.config.Config
 import org.apache.flink.api.common.restartstrategy.RestartStrategies
 import pl.touk.nussknacker.engine.api.async.{DefaultAsyncInterpretationValue, DefaultAsyncInterpretationValueDeterminer}
 import pl.touk.nussknacker.engine.api.namespaces.ObjectNaming
-import pl.touk.nussknacker.engine.api.process.{ProcessConfigCreator, ProcessObjectDependencies, RunMode}
+import pl.touk.nussknacker.engine.api.process.{ProcessConfigCreator, ProcessObjectDependencies, ComponentUsage}
 import pl.touk.nussknacker.engine.api.{JobData, MetaData, ProcessListener, ProcessVersion}
 import pl.touk.nussknacker.engine.compile._
 import pl.touk.nussknacker.engine.definition.DefinitionExtractor.ObjectWithMethodDef
@@ -34,13 +34,13 @@ class FlinkProcessCompiler(creator: ProcessConfigCreator,
                            val processConfig: Config,
                            val diskStateBackendSupport: Boolean,
                            objectNaming: ObjectNaming,
-                           val runMode: RunMode) extends Serializable {
+                           val componentUsage: ComponentUsage) extends Serializable {
 
   import net.ceedubs.ficus.Ficus._
   import net.ceedubs.ficus.readers.ArbitraryTypeReader._
   import pl.touk.nussknacker.engine.util.Implicits._
 
-  def this(modelData: ModelData) = this(modelData.configCreator, modelData.processConfig, diskStateBackendSupport = true, modelData.objectNaming, runMode = RunMode.Engine)
+  def this(modelData: ModelData) = this(modelData.configCreator, modelData.processConfig, diskStateBackendSupport = true, modelData.objectNaming, componentUsage = ComponentUsage.EngineRuntime)
 
   def compileProcess(process: EspProcess,
                      processVersion: ProcessVersion,
@@ -61,7 +61,7 @@ class FlinkProcessCompiler(creator: ProcessConfigCreator,
     val listenersToUse = listeners(processObjectDependencies)
 
     val compiledProcess =
-      ProcessCompilerData.prepare(process, definitions(processObjectDependencies), listenersToUse, userCodeClassLoader, resultCollector, runMode)
+      ProcessCompilerData.prepare(process, definitions(processObjectDependencies), listenersToUse, userCodeClassLoader, resultCollector, componentUsage)
 
     new FlinkProcessCompilerData(
       compiledProcess = compiledProcess,
@@ -70,7 +70,7 @@ class FlinkProcessCompiler(creator: ProcessConfigCreator,
       signalSenders = new FlinkProcessSignalSenderProvider(signalSenders(processObjectDependencies)),
       asyncExecutionContextPreparer = asyncExecutionContextPreparer,
       processTimeout = timeout,
-      runMode = runMode
+      componentUsage = componentUsage
     )
   }
 
@@ -91,14 +91,13 @@ class FlinkProcessCompiler(creator: ProcessConfigCreator,
                                  processObjectDependencies: ProcessObjectDependencies,
                                  listeners: Seq[ProcessListener],
                                  classLoader: ClassLoader): FlinkExceptionHandler = {
-    runMode match {
-      case _ : RunMode.Normal =>
-        new FlinkExceptionHandler(metaData, processObjectDependencies, listeners, classLoader)
-      case RunMode.Test =>
+    componentUsage match {
+      case ComponentUsage.TestRuntime =>
         new FlinkExceptionHandler(metaData, processObjectDependencies, listeners, classLoader) {
           override def restartStrategy: RestartStrategies.RestartStrategyConfiguration = RestartStrategies.noRestart()
           override def handle(exceptionInfo: NuExceptionInfo[_ <: Throwable]): Unit = ()
         }
+      case _ => new FlinkExceptionHandler(metaData, processObjectDependencies, listeners, classLoader)
     }
   }
 }