+++ engine/flink/executor/src/test/scala/pl/touk/nussknacker/engine/process/helpers/ProcessTestHelpers.scala
@@ -70,7 +70,7 @@ class ProcessBaseTestHelpers(data: List[SimpleRecord]) extends ProcessConfigCrea
     "eagerLifecycleService" -> WithCategories(EagerLifecycleService),
     "enricherWithOpenService" -> WithCategories(new EnricherWithOpenService),
     "serviceAcceptingOptionalValue" -> WithCategories(ServiceAcceptingScalaOption),
-    "returningRunModeService" -> WithCategories(ReturningRunModeService),
+    "returningComponentUsageService" -> WithCategories(ReturningComponentUsageService),
     "throwingNonTransientErrors" -> WithCategories(new ThrowingService(NonTransientException("test input", "test msg"))),
   )