+++ engine/flink/test-util/src/main/scala/pl/touk/nussknacker/engine/process/helpers/SampleNodes.scala
@@ -174,7 +174,7 @@ object SampleNodes {
                                   (implicit ec: ExecutionContext,
                                    collector: ServiceInvocationCollector,
                                    contextId: ContextId,
-                                   runMode: RunMode): Future[Any] = {
+                                   componentUsage: ComponentUsage): Future[Any] = {
           if (!opened) {
             throw new IllegalArgumentException
           }
@@ -195,7 +195,7 @@ object SampleNodes {
       override def invokeService(params: Map[String, Any])(implicit ec: ExecutionContext,
                                                            collector: ServiceInvocationCollector,
                                                            contextId: ContextId,
-                                                           runMode: RunMode): Future[Any] = {
+                                                           componentUsage: ComponentUsage): Future[Any] = {
         collector.collect(s"static-$static-dynamic-${params("dynamic")}", Option(())) {
           Future.successful(())
         }
@@ -215,7 +215,7 @@ object SampleNodes {
     @MethodToInvoke(returnType = classOf[SimpleRecordWithPreviousValue])
     def execute(@ParamName("stringVal") stringVal: String,
                 @ParamName("groupBy") groupBy: LazyParameter[String])
-               (implicit nodeId: NodeId, metaData: MetaData, runMode: RunMode) = FlinkCustomStreamTransformation((start: DataStream[Context], context: FlinkCustomNodeContext) => {
+               (implicit nodeId: NodeId, metaData: MetaData, componentUsage: ComponentUsage) = FlinkCustomStreamTransformation((start: DataStream[Context], context: FlinkCustomNodeContext) => {
       setUidToNodeIdIfNeed(context,
         start
           .flatMap(context.lazyParameterHelper.lazyMapFunction(groupBy))
@@ -353,7 +353,7 @@ object SampleNodes {
                                   (implicit ec: ExecutionContext,
                                    collector: ServiceInvocationCollector,
                                    contextId: ContextId,
-                                   runMode: RunMode): Future[Any] = {
+                                   componentUsage: ComponentUsage): Future[Any] = {
             val result = (1 to count)
               .map(_ => definition.asScala.map(_ -> params("toFill").asInstanceOf[String]).toMap)
               .map(TypedMap(_))
@@ -421,14 +421,14 @@ object SampleNodes {
 
   }
 
-  object TransformerAddingRunMode extends CustomStreamTransformer {
+  object TransformerAddingComponentUsage extends CustomStreamTransformer {
 
     @MethodToInvoke
     def execute = {
       FlinkCustomStreamTransformation((start: DataStream[Context], flinkCustomNodeContext: FlinkCustomNodeContext) => {
-        val runMode = flinkCustomNodeContext.runMode
+        val componentUsage = flinkCustomNodeContext.componentUsage
         start
-          .map(context => ValueWithContext[AnyRef](runMode, context))
+          .map(context => ValueWithContext[AnyRef](componentUsage, context))
       })
     }
 
@@ -679,7 +679,7 @@ object SampleNodes {
 
   object GenericParametersSink extends SinkFactory with SingleInputGenericNodeTransformation[Sink]  {
 
-    private val runModeDependency = TypedNodeDependency[RunMode]
+    private val componentUsageDependency = TypedNodeDependency[ComponentUsage]
 
     override type State = Nothing
 
@@ -709,14 +709,14 @@ object SampleNodes {
       override def prepareValue(dataStream: DataStream[Context], flinkNodeContext: FlinkCustomNodeContext): DataStream[ValueWithContext[Value]] = {
         dataStream
           .flatMap(flinkNodeContext.lazyParameterHelper.lazyMapFunction(params("value").asInstanceOf[LazyParameter[String]]))
-          .map((v: ValueWithContext[String]) => v.copy(value = s"${v.value}+$typ-$version+runMode:${runModeDependency.extract(dependencies)}"))
+          .map((v: ValueWithContext[String]) => v.copy(value = s"${v.value}+$typ-$version+componentUsage:${componentUsageDependency.extract(dependencies)}"))
       }
 
       override def registerSink(dataStream: DataStream[ValueWithContext[String]], flinkNodeContext: FlinkCustomNodeContext): DataStreamSink[_] =
         dataStream.map(_.value).addSink(SinkForStrings.toSinkFunction)
     }
 
-    override def nodeDependencies: List[NodeDependency] = List(runModeDependency)
+    override def nodeDependencies: List[NodeDependency] = List(componentUsageDependency)
   }
 
   object ProcessHelper {
@@ -768,7 +768,7 @@ object SampleNodes {
   object TypedJsonSource extends SourceFactory with ReturningType {
 
     @MethodToInvoke
-    def create(processMetaData: MetaData, runMode: RunMode, @ParamName("type") definition: java.util.Map[String, _]): Source = {
+    def create(processMetaData: MetaData, componentUsage: ComponentUsage, @ParamName("type") definition: java.util.Map[String, _]): Source = {
       new CollectionSource[TypedMap](new ExecutionConfig, List(), None, Typed[TypedMap]) with FlinkSourceTestSupport[TypedMap] with ReturningType {
 
         override def testDataParser: TestDataParser[TypedMap] = new EmptyLineSplittedTestDataParser[TypedMap] {
@@ -788,11 +788,11 @@ object SampleNodes {
 
   @JsonCodec case class KeyValue(key: String, value: Int, date: Long)
 
-  object ReturningRunModeService extends Service {
+  object ReturningComponentUsageService extends Service {
 
     @MethodToInvoke
-    def invoke(implicit runMode: RunMode): Future[RunMode] = {
-      Future.successful(runMode)
+    def invoke(implicit componentUsage: ComponentUsage): Future[ComponentUsage] = {
+      Future.successful(componentUsage)
     }
 
   }