+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/ProcessObjectDefinitionExtractor.scala
@@ -10,7 +10,7 @@ import scala.reflect.ClassTag
 class ProcessObjectDefinitionExtractor[F, T: ClassTag] extends AbstractMethodDefinitionExtractor[F] {
 
   override protected def expectedReturnType: Option[Class[_]] = Some(implicitly[ClassTag[T]].runtimeClass)
-  override protected def additionalDependencies: Set[Class[_]] = Set[Class[_]](classOf[MetaData], classOf[NodeId], classOf[RunMode])
+  override protected def additionalDependencies: Set[Class[_]] = Set[Class[_]](classOf[MetaData], classOf[NodeId], classOf[ComponentUsage])
 
 }